int_loop.sh: while => if conditional

+++ fuzzers/int_loop.sh
@@ -5,11 +5,14 @@ MAKEFLAGS=${MAKEFLAGS:-}
 echo "make: ${MAKE} ${MAKEFLAGS}"
 echo $MAKE
 mkdir -p todo;
-while
+while true; do
     ${MAKE} ${MAKEFLAGS} cleanprj;
     ${MAKE} ${MAKEFLAGS} build/todo.txt || exit 1;
-    test -s build/todo.txt;
-do
+    echo "Remaining: " $(wc -l build/todo_all.txt)
+    if [ \! -s build/todo.txt ] ; then
+        break
+    fi
+
     i=$((i+1));
     cp build/todo.txt todo/${i}.txt;
     cp build/todo_all.txt todo/${i}_all.txt;
